More performance improvements

+++ types/src/bitboard.rs
@@ -156,8 +156,8 @@ impl BitBoard {
         flipped_self
     }
 
-    pub fn trailing_zeros(&self) -> u32 {
-        self.0.trailing_zeros()
+    pub fn trailing_zeros(&self) -> u8 {
+        self.0.trailing_zeros() as u8
     }
 
     pub fn clear_lsb(&mut self) {
